using names for empty titles; moving files into common folders (validators, reconcilers)
