+++ src/pepr/operator/README.md
@@ -57,28 +57,28 @@ metadata:
 spec:
   exemptions:
     - policies:
-        - Disallow_Host_Namespaces
-        - Disallow_Privileged
-        - Require_Non_Root_User
-        - Drop_All_Capabilities
-        - Restrict_HostPath_Write
-        - Restrict_Volume_Types
+        - DisallowHostNamespaces
+        - DisallowPrivileged
+        - RequireNonRootUser
+        - DropAllCapabilities
+        - RestrictHostPathWrite
+        - RestrictVolumeTypes
       matcher:
         namespace: neuvector
         name: "^neuvector-enforcer-pod.*"
     
     - policies:
-        - Disallow_Privileged
-        - Require_Non_Root_User
-        - Drop_All_Capabilities
-        - Restrict_HostPath_Write
-        - Restrict_Volume_Types
+        - DisallowPrivileged
+        - RequireNonRootUser
+        - DropAllCapabilities
+        - RestrictHostPathWrite
+        - RestrictVolumeTypes
       matcher:
         namespace: neuvector
         name: "^neuvector-controller-pod.*"
 
     - policies:
-        - Drop_All_Capabilities
+        - DropAllCapabilities
       matcher:
         namespace: neuvector
         name: "^neuvector-prometheus-exporter-pod.*"  
@@ -87,23 +87,21 @@ spec:
 ### Key Files and Folders
 
 ```bash
-.
+src/pepr/operator/
 ├── controllers             # Core business logic called by the reconciler
-│   ├── exemptions          # Manages updating Pepr store with exemptions from UDS Exemption
-│   ├── istio               # Manages Istio VirtualServices and sidecar injection for UDS Packages/Namespace
-│   └── network             # Manages default and generated NetworkPolicies for UDS Packages/Namespace
+│   ├── exemptions          # Manages updating Pepr store with exemptions from UDS Exemption
+│   ├── istio               # Manages Istio VirtualServices and sidecar injection for UDS Packages/Namespace
+│   ├── keycloak            # Manages Keycloak client syncing
+│   └── network             # Manages default and generated NetworkPolicies for UDS Packages/Namespace
 ├── crd
-│   ├── generated           # Type files generated by `uds run -f src/pepr/tasks.yaml gen-crds`
-│   ├── sources             # CRD source files
-│   ├── exempt-validator.ts  # Validates UDS Exemption CRs with Pepr
+│   ├── generated           # Type files generated by `uds run -f src/pepr/tasks.yaml gen-crds`
+│   ├── sources             # CRD source files
 │   ├── migrate.ts          # Migrates older versions of UDS Package CRs to new version
 │   ├── register.ts         # Registers the UDS Package CRD with the Kubernetes API
-│   └── validator.ts        # Validates UDS Package CRs with Pepr
-├── enqueue.ts              # Serializes UDS Package CRs for processing by the reconciler
+│   └── validators          # Validates Custom Resources with Pepr
 ├── index.ts                # Entrypoint for the UDS Operator
-└── reconciler.ts           # Reconciles UDS Package CRs via the controllers
+└── reconcilers             # Reconciles Custom Resources via the controllers
 ```
-
 ### Flow
 
 The UDS Operator leverages a Pepr Watch. The following diagram shows the flow of the UDS Operator:
