+++ src/pepr/operator/crd/sources/exemption/v1alpha1.ts
@@ -77,7 +77,7 @@ export const v1alpha1: V1CustomResourceDefinitionVersion = {
                         "DisallowNodePortServices",
                         "DisallowPrivileged",
                         "DisallowSELinuxOptions",
-                        "Drop_AllCapabilities",
+                        "DropAllCapabilities",
                         "RequireNonRootUser",
                         "RestrictCapabilities",
                         "RestrictExternalNames",
