+++ src/pepr/operator/crd/validators/exempt-validator.spec.ts
@@ -1,9 +1,9 @@
 import { afterEach, describe, expect, it, jest } from "@jest/globals";
 import { PeprValidateRequest } from "pepr";
-import { MatcherKind, UDSExemption } from ".";
-import { UDSConfig } from "../../config";
+import { MatcherKind, UDSExemption } from "..";
+import { UDSConfig } from "../../../config";
+import { ExemptionElement, Policy } from "../generated/exemption-v1alpha1";
 import { exemptValidator } from "./exempt-validator";
-import { ExemptionElement, Policy } from "./generated/exemption-v1alpha1";
 
 type mockReqArgs = {
   ns?: string;
