+++ src/pepr/operator/index.ts
@@ -1,6 +1,6 @@
-import { Log, a } from "pepr";
+import { a } from "pepr";
 import { When } from "./common";
-import { processExemptions, removeExemptions } from "./controllers/exemptions/exemptions";
+import { removeExemptions } from "./controllers/exemptions/exemptions";
 import { cleanupNamespace } from "./controllers/istio/injection";
 import { purgeSSOClients } from "./controllers/keycloak/client-sync";
 import {
@@ -8,11 +8,12 @@ import {
   updateAPIServerCIDRFromEndpointSlice,
   updateAPIServerCIDRFromService,
 } from "./controllers/network/generators/kubeAPI";
-import { Phase, UDSExemption, UDSPackage } from "./crd";
-import { exemptValidator } from "./crd/exempt-validator";
+import { UDSExemption, UDSPackage } from "./crd";
 import "./crd/register";
-import { validator } from "./crd/validator";
-import { reconciler, updateStatus } from "./reconciler";
+import { exemptValidator } from "./crd/validators/exempt-validator";
+import { validator } from "./crd/validators/validator";
+import { exemptReconciler } from "./reconcilers/exempt-reconciler";
+import { reconciler } from "./reconcilers/reconciler";
 
 // Export the operator capability for registration in the root pepr.ts
 export { operator } from "./common";
@@ -58,42 +59,4 @@ When(UDSPackage)
 When(UDSExemption).IsDeleted().Watch(removeExemptions);
 
 // Watch for changes to the UDSExemption CRD to enqueue an exemption for processing
-When(UDSExemption)
-  .IsCreatedOrUpdated()
-  .Validate(exemptValidator)
-  .Reconcile(async (exempt: UDSExemption) => {
-    if (!exempt.metadata?.namespace) {
-      Log.error(exempt, `Invalid Exemption definition`);
-      return;
-    }
-
-    const isPending = exempt.status?.phase === Phase.Pending;
-    const isCurrentGeneration = exempt.metadata?.generation === exempt.status?.observedGeneration;
-
-    if (isPending || isCurrentGeneration) {
-      Log.debug(exempt, `Skipping pending or completed exemption`);
-      return;
-    }
-
-    const { namespace, name } = exempt.metadata;
-
-    Log.debug(exempt, `Processing Exemption ${namespace}/${name}`);
-
-    try {
-      await updateStatus(exempt, { phase: Phase.Pending });
-
-      processExemptions(exempt);
-      await updateStatus(exempt, {
-        phase: Phase.Ready,
-        observedGeneration: exempt.metadata.generation,
-        titles: exempt.spec?.exemptions?.map(e => e.title || ""),
-      });
-    } catch (e) {
-      Log.error(e, `Error configuring for ${namespace}/${name}`);
-      // todo: need to evaluate when it is safe to retry (updating generation now avoids retrying infinitely)
-      void updateStatus(exempt, {
-        phase: Phase.Failed,
-        observedGeneration: exempt.metadata.generation,
-      });
-    }
-  });
+When(UDSExemption).IsCreatedOrUpdated().Validate(exemptValidator).Reconcile(exemptReconciler);
